added some more debugging

+++ 52_maximus_parser/parse_gribs/utils/remove_directories.py
@@ -4,9 +4,11 @@ import os
 def removeDirectories(deleted_directory):
     while os.path.basename(deleted_directory) != "downloaded_grib_files":
         try:
-            os.rmdir(deleted_directory)
+            if os.path.exists(deleted_directory) and os.path.isdir(deleted_directory):
+                print(f"It exists: {deleted_directory}")
+                os.rmdir(deleted_directory)
         except OSError:
-            print(f"COntinue there?")
-            pass  # Continue the loop even if an OSError occurs
+            print("An error occurred while removing the directory.")
+
         deleted_directory = os.path.dirname(deleted_directory)
-        print(f"NExt deleted directory: {deleted_directory}")
+        print(f"Next deleted directory: {deleted_directory}")
